tests/requirements-test.txt
# Test dependencies for Command Center
pytest>=7.0.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
httpx>=0.24.0
numpy>=1.24.0
